Remove DecomposedContainerValueFactory
